backend/wrapper/pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "ensys"
version = "0.2a7"
dependencies = [
    'pydantic>=2.0',
    'oemof.solph==0.5.6',
    'pyrsistent'
]
authors=[
    {name = "Andreas Lubojanski", email = "andreas.lubojanski@hs-nordhausen.de"}
]
description = "EnSys is a backend to build energysystems from abstract configcontainers for oemof.solph."
license={text = "aGPL"}
requires-python=">=3.8"
classifiers=[
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent"
]

[tool.setuptools]
package-dir = {"" = "src"}
